backend/app/requirements.txt
flask
mysql-connector
googlemaps
PyJWT